Add leaning counters to local storage
Add three items in localStorage: left, right, and neutral. Upvoting an article will increase the reader's "score" for that leaning, and downvoting will decrese the "score".
For example, if a user reads an article by The New York Times and upvotes it, the "left" score will increase by one. As the reader moves through articles and topics, trends will
likely emerge in their reading, i.e. a highly negative or positive score in one or more categories: there is their bias.